Implementacion de Cache para busquedas

+++ Travel.Business/Entities/Journey.cs
@@ -9,7 +9,7 @@ namespace Travel.Business.Entities
     {
         public string Origin { get; set; }
         public string Destination { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int IdTypeRequest { get; set; }
         public List<Flight> Flights { get; set; }
 
@@ -17,7 +17,7 @@ namespace Travel.Business.Entities
         {
             public string Origin { get; set; }
             public string Destination { get; set; }
-            public double Price { get; set; }
+            public decimal Price { get; set; }
             public Transport Transport { get; set; }
         }
